// src/app/Model/Character.ts
export interface Character {
    id? :string;
    name?: string;
    HP?: number;
    XP?: number;
    player?: string;
    character?: string;
    canLvlUp?: boolean;
    rase?: string;
    signature?: string;
    strange?: number;
    dexterity?: number;
    condidtion?: number;
    wisdom?: number;
    intelligence?: number;
    charisma?: number;
    KP?: number;
    level?: number;


}